Seperated CMSIS-DAP layer from pyOCD
Before this commit, pyOCD seemed to be composed of different projects
with slightly different goals. The CMSIS-DAP layer that manages
communication to the DAP of attached devices and the collection of
board and flash configurations that handle on-chip debugging.

In an effort to better organize these projects, the CMSIS-DAP layer
has been split out into the pyDAPLink project and removed from pyOCD
except for a wrapper class similarly to how the interfaces were handled.